requirements.txt
asgiref==3.8.1
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
cssselect==1.2.0
d==0.2.2
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
gyp==0.1
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
Markdown==3.7
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
psycopg2-binary==2.9.10
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyquery==2.0.1
python-apt==2.4.0+ubuntu3
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.2
systemd-python==234
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
